+++ pkg/collector/loader/README.md
