+++ pkg/collector/loader/file_provider.go
