+++ pkg/collector/loader/file_provider_test.go
