+++ pkg/collector/loader/tests/invalid.yaml
