+++ pkg/collector/loader/tests/notaconfig.yaml
