+++ pkg/collector/loader/tests/testcheck.yaml
