+++ pkg/collector/loader/types.go
